fix Xcode compilation issue

+++ adapters/tlsio_openssl.c
@@ -1821,7 +1821,7 @@ static int setup_crl_check(TLS_IO_INSTANCE* tls_io_instance)
     store = SSL_CTX_get_cert_store(tls_io_instance->ssl_context);
 
 #if USE_OPENSSL_1_1_0_OR_UP
-    int flags = X509_VERIFY_PARAM_get_flags(X509_STORE_get0_param(store));
+    long flags = X509_VERIFY_PARAM_get_flags(X509_STORE_get0_param(store));
 #else
     long flags = X509_VERIFY_PARAM_get_flags(store->param);
 #endif
